Refactor of Vault implementation
